API_AI/Chatbot/requirements.txt
gradio==3.14
googlesearch-python==1.1.0
pyodbc
requests
